lab/04_metadata_rag_mart/requirements.txt
psycopg2-binary>=2.9
openai>=1.30
voyageai>=0.3.0
anthropic>=0.39
ollama>=0.4
pandas>=2.0
pyarrow>=15.0
boto3>=1.34
pydantic>=2.0
streamlit>=1.40
plotly>=5.20
scikit-learn>=1.4
